+++ liveviz/src/actions/GlobalActions.js
@@ -11,9 +11,3 @@ export const setConnectionStatus = (status) => {
     val: status,
   }
 }
-
-export const setParamsStatus = () => {
-    return {
-        type: 'SET_PARAMS_STATUS'
-    }
-}
