+++ liveviz/src/reducers/GlobalReducers.js
@@ -4,7 +4,6 @@ const defaultState = {
   serverRoot: 'http://209.9.36.2:1234',
   connectionStatus: 'connected',
   graphKey: 0,
-  isParamEditable: false,
 }
 
 export const global = (state=defaultState, action) => {
@@ -13,8 +12,6 @@ export const global = (state=defaultState, action) => {
     return merge(state, {serverRoot: action.val, graphKey: state.graphKey + 1});
     case 'SET_CONNECTION_STATUS':
     return assoc('connectionStatus', action.val, state);
-    case 'SET_PARAMS_STATUS':
-    return merge(state, {isParamEditable : !state.isParamEditable})
     default:
     return state
   }
